SEO Optimization Phase 1.1: Candidacy Check Page Overhaul & Progress Tracking
✅ Completed:
- /find-care/candidacy-check → converted to server component
  • Optimized <h1>–<h4> hierarchy (1 <h1>, semantic steps)
  • Added metadata (title & description) for SEO
  • Improved accessibility and search clarity

🛠️ SEO Fixes Completed So Far:
- Homepage
- All Treatments Pages
- All Condition Pages
- All Doctor Profile Pages
- Blog Listing Page
- Single Blog Pages
- /about
- /find-care/book-an-appointment
- /about/faqs
- /condition-check
- /find-care/candidacy-check ✅

Remaining Pages:
- /find-care/free-mri-review
- /patient-forms
- /faqs

- /area-of-speciality/faqs
