+++ app/blogs/layout.tsx
@@ -17,6 +17,6 @@ export const metadata = {
   },
 };
 
-export default function Layout({ children }) {
+export default function Layout({ children }: { children: React.ReactNode }) {
   return <>{children}</>;
 }
